requirements.txt
Django==2.2.1
pytz==2019.1
sqlparse==0.3.0
djangorestframework
djangorestframework-jwt
django-rest-swagger